Fade on checked in ShopList, clip name to length
The textviews in Shopping List and Pantry activities now truncate names
appropriately such that there is only one line of text with an ellipsis
at the end if it runs too long.

On "checked off" in the Shopping List activity, items' text now fades to
indicate this action more visually.